+++ src/extension.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 import getWebviewContent from "./iframeView";
 import * as path from "path";
-import { VsChatSidebarProvider } from "./VsChatSidebarProvider";
+import { VsChatSidebarProvider } from "./vsChatSidebarProvider";
 
 export function activate(context: vscode.ExtensionContext) {
   const vsChatSidebarProvider = new VsChatSidebarProvider(context.extensionUri);