+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to the "vschat" extension will be documented in this file.
 
+## [0.0.5] - 2020-11-30
+- Add Chatroom list: users can now select the lobby
+- Bonus: Add Snow to chatroom (Happy New Year!)
+- Add Loading screen to remove iframe flicker
+- Update Chatroom Theme (thanks to @steffan and @lxhom)
+- Fix Retaining Context When Hidden
+
+## [0.0.4] - 2020-11-28
+- Fix Chatroom Theme
+- Update README.md
+
+## [0.0.3] - 2020-11-27
+- Fix some bugs
+
 ## [0.0.2] - 2020-11-26
 - Add Demo to README.md